+++ packages/core/schematics/ng-generate/control-flow-migration/index.ts
@@ -13,8 +13,11 @@ import {normalizePath} from '../../utils/change_tracker';
 import {getProjectTsConfigPaths} from '../../utils/project_tsconfig_paths';
 import {canMigrateFile, createMigrationProgram} from '../../utils/typescript/compiler_host';
 
+import {migrateFor} from './fors';
+import {migrateIf} from './ifs';
+import {migrateSwitch} from './switches';
 import {AnalyzedFile, MigrateError} from './types';
-import {analyze, migrateTemplate} from './util';
+import {analyze, processNgTemplates} from './util';
 
 interface Options {
   path: string;
@@ -84,7 +87,18 @@ function runControlFlowMigration(
     for (const [start, end] of ranges) {
       const template = content.slice(start, end);
       const length = (end ?? content.length) - start;
-      const {migrated, errors} = migrateTemplate(template);
+
+      const ifResult = migrateIf(template);
+      const forResult = migrateFor(ifResult.migrated);
+      const switchResult = migrateSwitch(forResult.migrated);
+
+      const errors = [
+        ...ifResult.errors,
+        ...forResult.errors,
+        ...switchResult.errors,
+      ];
+
+      const migrated = processNgTemplates(switchResult.migrated);
 
       if (migrated !== null) {
         update.remove(start, length);
